new feature: unnamed buffer
When called with '*' as parameter, notevim now opens a new, unnamed
file. This allows notevim to be used not only for permanent notes,
but also for temporary buffers that will not be of interest later on.
Of course you can still save them and have everything work as expected.

The dmenu and openbox menus also list this option of course.